pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "hydrodataset"
dynamic = ["version"]
description = "A Python package for downloading and reading hydrological datasets"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Wenyu Ouyang", email = "wenyuouyang@outlook.com"}
]
maintainers = [
    {name = "Wenyu Ouyang", email = "wenyuouyang@outlook.com"}
]
keywords = ["hydrodataset", "hydrology", "datasets", "water"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "xarray",
    "netCDF4",
    "zarr",
    "dask[complete]",
    "dask[distributed]",
    "bokeh>=2.4.2",
    "scipy",
    "pint",
    "pint-pandas",
    "pint-xarray",
    "tzfpy",
    "codetiming",
    "hydroutils",
    "openpyxl>=3.1.5",
    "xlrd>=2.0.2",
    "requests>=2.32.4",
    "py7zr>=1.0.0",
    "aqua-fetch[all]",
]

[project.optional-dependencies]
dev = [
    "bump2version",
    "wheel",
    "watchdog",
    "tox",
    "coverage",
    "Sphinx",
    "twine",
    "grip",
    "black",
    "flake8",
    "pytest>=3",
    "pytest-runner",
]
docs = [
    "Sphinx",
    "grip",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-git-revision-date-plugin",
    "mkdocs-jupyter",
    "ipykernel",
    "mkdocstrings-python",
]
lint = [
    "black",
    "flake8",
]
release = [
    "build",
    "twine",
    "hatchling",
    "hatch-vcs",
]

[project.urls]
Homepage = "https://github.com/OuyangWenyu/hydrodataset"
Repository = "https://github.com/OuyangWenyu/hydrodataset"
Documentation = "https://github.com/OuyangWenyu/hydrodataset"
"Bug Tracker" = "https://github.com/OuyangWenyu/hydrodataset/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["hydrodataset"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "docs/source/conf.py",
    "old",
    "build",
    "dist",
    ".eggs",
    "*.egg",
]


# Hatch version management
[tool.hatch.version]
source = "vcs"
fallback-version = "0.2.0"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[tool.hatch.build.hooks.vcs]
version-file = "hydrodataset/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["hydrodataset"]

[tool.hatch.build.targets.sdist]
include = [
    "/hydrodataset",
    "/tests",
    "/docs",
    "/README.md",
    "/LICENSE",
    "/AUTHORS.rst",
    "/MANIFEST.in",
]

[tool.hatch.metadata]
allow-direct-references = true

# UV-specific configuration
[tool.uv]
# Override dependencies from aqua-fetch to use more flexible versions
override-dependencies = [
    "pandas>=2.2.0",           # Override aqua-fetch's pandas==2.1.4
]
